fix: refactor util functions to allow specific imports in Cypress
