+++ cypress/e2e/folders.spec.ts
@@ -1,10 +1,7 @@
 import slugify from "slugify"
 
-import {
-  titleToPageFileName,
-  pageFileNameToTitle,
-  deslugifyDirectory,
-} from "utils"
+import { deslugifyDirectory } from "utils/deslugify"
+import { titleToPageFileName, pageFileNameToTitle } from "utils/legacyUtils"
 
 import {
   CMS_BASEURL,
