+++ cypress/e2e/resources.spec.ts
@@ -1,4 +1,4 @@
-import { slugifyCategory } from "utils"
+import { slugifyCategory } from "utils/legacyUtils"
 
 import {
   CMS_BASEURL,
