+++ src/utils/index.ts
@@ -8,6 +8,7 @@ export * from "./siteColorUtils"
 export * from "./toasts"
 export * from "./validators"
 export * from "./legacy"
+export * from "./legacyUtils"
 export * from "./text"
 export * from "./date"
 export * from "./files"
